refactor: split the results-filter index tests by subject
The module reached 1558 lines, over the 1000-line cap, and it had grown
four subjects: the parity matrix that holds the two implementations to
one answer, the enumerated divergences, what the filter reports about
the pass it is answering from, and the refusals.

The fixture tree and the helpers that build it move to a package
conftest, losing their leading underscores because they are now
imported; the tests move whole. The collected test identifiers are
byte-identical before and after, and every mutation that pinned a test
before the move still fails exactly the same test.
